add_replay_synapses & add_replay_minis to use same syn location calculation
